Optional delelte all

+++ src/main/java/com/btb/sne/batch/JobLoggerListener.java
@@ -1,5 +1,6 @@
 package com.btb.sne.batch;
 
+import com.btb.sne.config.ApplicationConfig;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.batch.core.JobExecution;
@@ -15,12 +16,15 @@ public class JobLoggerListener {
     private static final String START_MESSAGE = "%s is beginning execution";
     private static final String END_MESSAGE = "%s has completed with the status %s";
 
+    private final ApplicationConfig config;
     private final InitService initService;
 
     @BeforeJob
     public void beforeJob(JobExecution jobExecution) {
         log.info(String.format(START_MESSAGE, jobExecution.getJobInstance().getJobName()));
-        initService.deleteAll();
+        if (config.isDelete()) {
+            initService.deleteAll();
+        }
     }
 
     @AfterJob
